Rename parsing and transliteration errors to failures
The convention from this point onward will be to reserve the term error
for unexpected events (eg: failing to encode a HangeulSyllabicBlock
instance to a unicode character) while the term failure will be used for
usually undesirable but expected events (eg: failing to parse a string
containing both hangeul syllables and other characters as a single
HangeulTextElement instance). The top-level type remains called Error
though.